Add a proper cast.

+++ src/boson/src/internal/thread.cc
@@ -51,7 +51,7 @@ void thread::unregister_all_events() {
   loop_.unregister(self_event_id_);
 }
 
-thread::thread(engine& parent_engine) : engine_proxy_(parent_engine), loop_(*this), engine_queue_{parent_engine.max_nb_cores()+1} {
+thread::thread(engine& parent_engine) : engine_proxy_(parent_engine), loop_(*this), engine_queue_{static_cast<int>(parent_engine.max_nb_cores()+1)} {
   engine_event_id_ = loop_.register_event(&engine_event_id_);
   self_event_id_ = loop_.register_event(&self_event_id_);
   engine_proxy_.set_id();  // Tells the engine which thread id we got
